drying out unit tests
